fix: address PR #559 round-2 test/doc nits
Low (all explicitly non-blocking):
- docs/dispatch.md:456 — table row referenced `gate::classify`; missed
  in the round-1 sweep. Now points at `dispatch_gate::classify` with
  the §9 #9 rename callout.
- sipag-lens BridgeChatBackend::chat_with_options — dropped the
  unnecessary `t as f64` cast. `serde_json::Value::from(f32)` exists
  via the From impl; the explicit widening was harmless but produced
  no observable difference in the JSON output.

Deferred (explicitly per round-2 reviewer):
- No unit test for OLLAMA_MODEL env-override. `std::env::set_var` is
  process-global and races with parallel-test reads of the same var;
  the inline source comment explains the defer. The default-when-unset
  path is exercised by the existing routing test.

Validation: 12 dispatch_gate tests still pass; sipag-lens tests still
pass; cargo fmt + clippy -D warnings + workspace test all green.

+++ sipag-lens/src/lib.rs
@@ -662,7 +662,7 @@ impl ChatBackend for BridgeChatBackend {
         // body (identical to the plain `chat_messages` path).
         let mut options_obj = serde_json::Map::new();
         if let Some(t) = options.temperature {
-            options_obj.insert("temperature".into(), serde_json::Value::from(t as f64));
+            options_obj.insert("temperature".into(), serde_json::Value::from(t));
         }
         if let Some(n) = options.num_predict {
             options_obj.insert("num_predict".into(), serde_json::Value::from(n));
